Homework после ревью2

+++ homework.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from typing import list
 
 
 @dataclass
@@ -101,7 +102,7 @@ class Swimming(Training):
         return calories * self.COEF_CAL_7 * self.weight
 
 
-def read_package(workout_type: str, data: list[int]) -> Training:
+def read_package(workout_type: str, data: list[float]) -> Training:
     """Прочитать данные полученные от датчиков."""
     type_dict = {'SWM': Swimming, 'RUN': Running, 'WLK': SportsWalking}
     return type_dict[workout_type](*data)
